clean up

+++ L02/push_swap/libft/ft_calloc.c
@@ -16,7 +16,7 @@ void	*ft_calloc(size_t nmemb, size_t size)
 {
 	void	*p;
 
-	if (size != 0 && nmemb > SIZE_MAX / size)
+	if (size != 0 && nmemb > INT_MAX / size)
 	{
 		return (NULL);
 	}
